+++ provider.tf
@@ -18,6 +18,11 @@ provider "azurerm" {
     key_vault {
       purge_soft_delete_on_destroy    = false
       recover_soft_deleted_key_vaults = false
+      recover_soft_deleted_keys       = false
+      recover_soft_deleted_secrets    = false
+    }
+    resource_group {
+      prevent_deletion_if_contains_resources = false
     }
   }
 }