+++ variables.tf
@@ -2,10 +2,12 @@
 # Azure basic vars
 #-----------------------
 variable "rg_name" {
+  type    = string
   default = "vault-rg"
 }
 
 variable "location" {
+  type    = string
   default = "canadacentral"
 }
 
@@ -16,6 +18,7 @@ variable "location" {
 
 variable "acr_name" {
   description = "Name of Azure Container Registry repo. Alphanumeric characters only without spaces."
+  type        = string
   default     = "vaultacr"
 }
 
@@ -25,16 +28,19 @@ variable "acr_name" {
 #-------------------------
 variable "storage_account_name" {
   description = "Name of Azure storage account. Lowercase alphanumeric characters only."
+  type        = string
   default     = "vaultstorage"
 }
 
 variable "storage_share_name" {
   description = "Name of storage share used to hold the vault-server.hcl to be mounted onto ACI."
+  type        = string
   default     = "vault-config"
 }
 
 variable "storage_container_name" {
   description = "Name of storage container used as Vault's storage backend."
+  type        = string
   default     = "vault-data"
 }
 
@@ -44,10 +50,12 @@ variable "storage_container_name" {
 #-------------------------
 variable "akv_name" {
   description = "Name of AKV used to store key used for Auto-unseal."
+  type        = string
   default     = "vault-akv"
 }
 
 variable "soft_delete_retention_days" {
+  type    = number
   default = 7
 }
 
@@ -56,9 +64,11 @@ variable "soft_delete_retention_days" {
 # HashiCorp Vault 
 #-------------------------
 variable "vault_name" {
+  type    = string
   default = "azvault"
 }
 
 variable "vault_version" {
+  type    = string
   default = "1.10.4"
 }